Moved code from pages to components, simplified pages code.
